// src/gen/circle/v1/circle_connect.ts
// @generated by protoc-gen-connect-es v1.7.0 with parameter "target=ts"
// @generated from file circle/v1/circle.proto (package circles.v1.circle, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { Circle, CreateCircleRequest, GetCircleRequest, ListCirclesRequest, ListCirclesResponse, UpdateCircleStatusRequest } from "./circle_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service circles.v1.circle.CircleService
 */
export const CircleService = {
  typeName: "circles.v1.circle.CircleService",
  methods: {
    /**
     * @generated from rpc circles.v1.circle.CircleService.CreateCircle
     */
    createCircle: {
      name: "CreateCircle",
      I: CreateCircleRequest,
      O: Circle,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc circles.v1.circle.CircleService.GetCircle
     */
    getCircle: {
      name: "GetCircle",
      I: GetCircleRequest,
      O: Circle,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc circles.v1.circle.CircleService.ListCircles
     */
    listCircles: {
      name: "ListCircles",
      I: ListCirclesRequest,
      O: ListCirclesResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc circles.v1.circle.CircleService.UpdateCircleStatus
     */
    updateCircleStatus: {
      name: "UpdateCircleStatus",
      I: UpdateCircleStatusRequest,
      O: Circle,
      kind: MethodKind.Unary,
    },
  }
} as const;